+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+### KiteUtils v0.7.1 - 2024-07-24
+#### Added
+- new parameters `winch_model`, `drum_radius`, `gear_ratio`, `inertia_motor`
+
 ### KiteUtils v0.6.16 - 2024-06-25
 #### Changed
 - new field `width_3l`